Split CLI into multiple files
